removed useless css from PasswordField and TextField, refactored scss styles in the components
