+++ src/components/Fields/PasswordField/PasswordField.module.scss
@@ -1,53 +1,57 @@
 .text-field {
   @apply mb-4;
-}
 
-.text-field--has-error .text-field__label-text {
-  @apply text-red-1130;
-}
+  &--has-error .text-field__label-text {
+    @apply text-red-1130;
+  }
 
-.text-field--has-error .text-field__input {
-  @apply border-red-1130;
-}
+  &--has-error .text-field__input {
+    @apply border-red-1130;
+  }
 
-.text-field__label {
-  @apply block;
-}
+  &__label {
+    @apply block;
+  }
 
-.text-field__label-text {
-  @apply text-gray-1030 text-xs;
-}
+  &__label-text {
+    @apply text-gray-1030 text-xs;
+  }
 
-.text-field__holder {
-  @apply relative;
-}
+  &__holder {
+    @apply relative;
+  }
 
-.text-field__input {
-  @apply text-gray-1030 text-sm h-9 py-2 px-2.5 outline-none border-gray-1030 border-2 rounded-md w-full;
-}
+  &__input {
+    @apply text-gray-1030 text-sm h-9 py-2 pl-2.5 pr-8 outline-none border-gray-1030 border-2 rounded-md w-full;
+  }
 
-.text-field__error-message {
-  @apply text-red-1130 text-2xs mt-1;
-}
+  &__icon {
+    @apply absolute top-1/2 right-[12px] translate-y-[-50%] cursor-pointer;
+  }
 
-.text-field__requirements {
-  @apply text-gray-1030 text-2xs mt-1;
+  &__error-message {
+    @apply text-red-1130 text-2xs mt-1;
+  }
+
+  &__requirements {
+    @apply text-gray-1030 text-2xs mt-1;
+  }
 }
 
 .strength-progress {
   @apply mt-2 grid grid-cols-3 gap-4;
-}
 
-.strength-progress--weak .strength-progress__item--active {
-  @apply bg-red-1130;
-}
-.strength-progress--okey .strength-progress__item--active {
-  @apply bg-[#ffbd5a];
-}
-.strength-progress--strong .strength-progress__item--active {
-  @apply bg-green-1010;
-}
+  &--weak .strength-progress__item--active {
+    @apply bg-red-1130;
+  }
+  &--okey .strength-progress__item--active {
+    @apply bg-[#ffbd5a];
+  }
+  &--strong .strength-progress__item--active {
+    @apply bg-green-1010;
+  }
 
-.strength-progress__item {
-  @apply h-2 rounded-md bg-[#d9d9d9];
+  &__item {
+    @apply h-2 rounded-md bg-[#d9d9d9];
+  }
 }
