--- src/components/Fields/TextField/TextField.module.css
@@ -1,38 +1,0 @@
-.text-field {
-  @apply mb-4;
-}
-
-.text-field--success .text-field__input {
-}
-
-.text-field--has-error .text-field__label-text {
-  @apply text-red-1130;
-}
-
-.text-field--has-error .text-field__input {
-  @apply border-red-1130;
-}
-
-.text-field__label {
-  @apply block;
-}
-
-.text-field__label-text {
-  @apply text-gray-1030 text-xs;
-}
-
-.text-field__holder {
-  @apply relative;
-}
-
-.text-field__input {
-  @apply text-gray-1030 text-sm h-9 py-2 pl-2.5 pr-8 outline-none border-gray-1030 border-2 rounded-md w-full;
-}
-
-.text-field__icon {
-  @apply absolute top-1/2 right-[12px] translate-y-[-50%] cursor-pointer;
-}
-
-.text-field__error-message {
-  @apply text-red-1130 text-2xs mt-1;
-}
