+++ src/components/Fields/TextField/TextField.module.scss
@@ -1,31 +1,38 @@
 .text-field {
   @apply mb-4;
+
+  &--has-error .text-field__label-text {
+    @apply text-red-1130;
+  }
+
+  &--has-error .text-field__input {
+    @apply border-red-1130;
+  }
+
+  &__label {
+    @apply block;
+  }
+
+  &__label-text {
+    @apply text-gray-1030 text-xs;
+  }
+
+  &__holder {
+    @apply relative;
+  }
+
+  &__input {
+    @apply text-gray-1030 text-sm h-9 py-2 pl-2.5 pr-8 outline-none border-gray-1030 border-2 rounded-md w-full;
+  }
+
+  &__icon {
+    @apply absolute top-1/2 right-[12px] translate-y-[-50%] cursor-pointer;
+  }
+
+  &__error-message {
+    @apply text-red-1130 text-2xs mt-1;
+  }
 }
 
-.text-field--has-error .text-field__label-text {
-  @apply text-red-1130;
-}
-
-.text-field--has-error .text-field__input {
-  @apply border-red-1130;
-}
-
-.text-field__label {
-  @apply block;
-}
-
-.text-field__label-text {
-  @apply text-gray-1030 text-xs;
-}
-
-.text-field__holder {
-  @apply relative;
-}
-
-.text-field__input {
-  @apply text-gray-1030 text-sm h-9 py-2 px-2.5 outline-none border-gray-1030 border-2 rounded-md w-full;
-}
-
-.text-field__error-message {
-  @apply text-red-1130 text-2xs mt-1;
+.text-field--success {
 }
